Standardize service profiles and package layout
