+++ networking-monorepo/docker/compose/dev.yml
@@ -58,7 +58,7 @@ services:
       GITHUB_CLIENT_ID: ${GITHUB_CLIENT_ID}
       GITHUB_CLIENT_SECRET: ${GITHUB_CLIENT_SECRET}
       APP_SECURITY_ISSUER_URI: http://identity-provider:8081
-    command: sh -c "cd services/identity-provider && ./gradlew --project-cache-dir /tmp/gradle-project-cache bootRun"
+    command: sh -c "cd services/identity-provider && ./gradlew --project-cache-dir /tmp/gradle-project-cache bootRunInContainer"
 
   app-service:
     container_name: app-service
@@ -67,7 +67,7 @@ services:
     ports:
       - "8082:8080"
       - "5002:5005"
-    command: sh -c "cd services/app-service && ../../gradlew --project-cache-dir /tmp/gradle-project-cache bootRun"
+    command: sh -c "cd services/app-service && ./gradlew --project-cache-dir /tmp/gradle-project-cache bootRunInContainer"
 
   channel-service:
     container_name: channel-service
@@ -76,7 +76,7 @@ services:
     ports:
       - "8083:8080"
       - "5003:5005"
-    command: sh -c "cd services/channel-service && ../../gradlew --project-cache-dir /tmp/gradle-project-cache bootRun"
+    command: sh -c "cd services/channel-service && ./gradlew --project-cache-dir /tmp/gradle-project-cache bootRunInContainer"
 
   insights-service:
     container_name: insights-service
@@ -85,7 +85,7 @@ services:
     ports:
       - "8084:8080"
       - "5004:5005"
-    command: sh -c "cd services/insights-service && ../../gradlew --project-cache-dir /tmp/gradle-project-cache bootRun"
+    command: sh -c "cd services/insights-service && ./gradlew --project-cache-dir /tmp/gradle-project-cache bootRunInContainer"
 
   ai-layer:
     container_name: ai-layer
@@ -94,7 +94,7 @@ services:
     ports:
       - "8085:8080"
       - "5005:5005"
-    command: sh -c "cd services/ai-layer && ../../gradlew --project-cache-dir /tmp/gradle-project-cache bootRun"
+    command: sh -c "cd services/ai-layer && ./gradlew --project-cache-dir /tmp/gradle-project-cache bootRunInContainer"
 
   gateway:
     container_name: gateway
@@ -107,7 +107,7 @@ services:
       CONSUL_HOST: consul
       CONSUL_PORT: 8500
       IDP_ISSUER_URI: http://identity-provider:8081
-    command: sh -c "cd services/gateway && ../../gradlew --project-cache-dir /tmp/gradle-project-cache bootRun"
+    command: sh -c "cd services/gateway && ./gradlew --project-cache-dir /tmp/gradle-project-cache bootRunInContainer"
 
 volumes:
   pnpm-store:
