--- networking-monorepo/services/ai-layer/src/main/java/com/example/ailayer/AiLayerApplication.java
@@ -1,11 +1,0 @@
-package com.example.ailayer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AiLayerApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(AiLayerApplication.class, args);
-    }
-}
