+++ networking-monorepo/services/ai-layer/src/main/java/ocudni104/ailayer/AiLayerApplication.java
@@ -1,0 +1,11 @@
+package ocudni104.ailayer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AiLayerApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(AiLayerApplication.class, args);
+    }
+}
