+++ networking-monorepo/services/ai-layer/src/main/resources/application.yml
@@ -1,0 +1,12 @@
+server:
+  port: 8080
+
+spring:
+  application:
+    name: ai-layer
+
+management:
+  endpoints:
+    web:
+      exposure:
+        include: health,info
