--- networking-monorepo/services/app-service/src/docker/workspace-manager.yml
@@ -1,11 +1,0 @@
-services:
-  workspace-manager:
-    build: ./../
-    ports:
-      - "8081:8081"
-      - "5005:5005"
-    environment:
-      - JAVA_TOOL_OPTIONS=-agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=*:5005
-    # EL COMANDO CLAVE: Arranca desde clases, NO desde un jar
-    command: >
-      java -cp /app/resources:/app/classes:/app/libs/* com.example.Main
