--- networking-monorepo/services/app-service/src/main/java/com/example/Main.java
@@ -1,13 +1,0 @@
-package com.example;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Main {
-
-    public static void main(String[] args) {
-        // Esta línea arranca todo el contexto de Spring
-        SpringApplication.run(Main.class, args);
-    }
-}
