+++ networking-monorepo/services/app-service/src/main/java/ocudni104/appservice/Main.java
@@ -1,0 +1,13 @@
+package ocudni104.appservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Main {
+
+    public static void main(String[] args) {
+        // Esta línea arranca todo el contexto de Spring
+        SpringApplication.run(Main.class, args);
+    }
+}
