--- networking-monorepo/services/app-service/src/main/resources/application.properties
@@ -1,19 +1,0 @@
-server.port=8080
-spring.application.name=workspace-manager
-spring.cloud.consul.host=${CONSUL_HOST:localhost}
-spring.cloud.consul.port=${CONSUL_PORT:8500}
-spring.cloud.consul.discovery.service-name=app-service
-spring.cloud.consul.discovery.prefer-ip-address=true
-spring.cloud.consul.discovery.health-check-path=/actuator/health
-management.endpoints.web.exposure.include=health,info
-
-
-logging.level.org.springframework.web=INFO
-logging.level.com.example.nm-workspace-manager=DEBUG
-
-# spring.datasource.url=jdbc:mysql://localhost:3306/basedb
-# spring.datasource.username=root
-# spring.datasource.password=password
-# spring.jpa.hibernate.ddl-auto=update
-
-spring.jpa.show-sql=true
