--- networking-monorepo/services/channel-service/src/main/java/com/example/channel/ChannelServiceApplication.java
@@ -1,11 +1,0 @@
-package com.example.channel;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ChannelServiceApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(ChannelServiceApplication.class, args);
-    }
-}
