+++ networking-monorepo/services/channel-service/src/main/java/ocudni104/channel/ChannelServiceApplication.java
@@ -1,0 +1,11 @@
+package ocudni104.channel;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ChannelServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(ChannelServiceApplication.class, args);
+    }
+}
