+++ networking-monorepo/services/channel-service/src/main/resources/application-container.yml
@@ -1,0 +1,9 @@
+spring:
+  cloud:
+    consul:
+      host: ${CONSUL_HOST:localhost}
+      port: ${CONSUL_PORT:8500}
+      discovery:
+        service-name: channel-service
+        prefer-ip-address: true
+        health-check-path: /actuator/health
