+++ networking-monorepo/services/channel-service/src/main/resources/application-local.yml
@@ -1,0 +1,6 @@
+spring:
+  cloud:
+    consul:
+      discovery:
+        enabled: false
+        register: false
