--- networking-monorepo/services/channel-service/src/main/resources/application.properties
@@ -1,8 +1,0 @@
-server.port=8080
-spring.application.name=channel-service
-spring.cloud.consul.host=${CONSUL_HOST:localhost}
-spring.cloud.consul.port=${CONSUL_PORT:8500}
-spring.cloud.consul.discovery.service-name=channel-service
-spring.cloud.consul.discovery.prefer-ip-address=true
-spring.cloud.consul.discovery.health-check-path=/actuator/health
-management.endpoints.web.exposure.include=health,info
