+++ networking-monorepo/services/channel-service/src/main/resources/application.yml
@@ -1,0 +1,12 @@
+server:
+  port: 8080
+
+spring:
+  application:
+    name: channel-service
+
+management:
+  endpoints:
+    web:
+      exposure:
+        include: health,info
