--- networking-monorepo/services/gateway/src/main/java/ng/elagent/gateway/GatewayApplication.java
@@ -1,11 +1,0 @@
-package ng.elagent.gateway;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GatewayApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(GatewayApplication.class, args);
-    }
-}
