+++ networking-monorepo/services/gateway/src/main/java/ocudni104/gateway/GatewayApplication.java
@@ -1,0 +1,11 @@
+package ocudni104.gateway;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GatewayApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(GatewayApplication.class, args);
+    }
+}
