+++ networking-monorepo/services/gateway/src/main/resources/application-container.yml
@@ -1,0 +1,9 @@
+spring:
+  cloud:
+    consul:
+      host: ${CONSUL_HOST:localhost}
+      port: ${CONSUL_PORT:8500}
+      discovery:
+        service-name: gateway
+        prefer-ip-address: true
+        health-check-path: /actuator/health
