+++ networking-monorepo/services/gateway/src/main/resources/application.yml
@@ -6,19 +6,12 @@ spring:
     web-application-type: reactive
   application:
     name: nm-api-gateway
-  cloud:
-    consul:
-      host: ${CONSUL_HOST:localhost}
-      port: ${CONSUL_PORT:8500}
-      discovery:
-        service-name: gateway
-        prefer-ip-address: true
-        health-check-path: /actuator/health
   security:
     oauth2:
       resourceserver:
         jwt:
           issuer-uri: ${IDP_ISSUER_URI:http://localhost:8081}
+  cloud:
     gateway:
       discovery:
         locator:
