--- networking-monorepo/services/identity-provider/src/main/java/ng/elagent/idp/IdpApplication.java
@@ -1,11 +1,0 @@
-package ng.elagent.idp;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class IdpApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(IdpApplication.class, args);
-    }
-}
