--- networking-monorepo/services/identity-provider/src/main/java/ng/elagent/idp/config/SecurityConfig.java
@@ -1,33 +1,0 @@
-package ng.elagent.idp.config;
-
-import ng.elagent.idp.federation.FederatedIdentityAuthenticationSuccessHandler;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.annotation.Order;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-public class SecurityConfig {
-
-    @Bean
-    @Order(2)
-    public SecurityFilterChain defaultSecurityFilterChain(HttpSecurity http) throws Exception {
-        http
-            .authorizeHttpRequests(auth -> auth
-                .requestMatchers(
-                    "/.well-known/openid-configuration",
-                    "/oauth2/jwks",
-                    "/actuator/health",
-                    "/error"
-                ).permitAll()
-                .anyRequest().authenticated()
-            )
-            .oauth2Login(oauth2 -> oauth2
-                .successHandler(new FederatedIdentityAuthenticationSuccessHandler())
-            )
-            .formLogin(form -> form.loginPage("/login").permitAll());
-
-        return http.build();
-    }
-}
