+++ networking-monorepo/services/identity-provider/src/main/java/ocudni104/idp/IdpApplication.java
@@ -1,0 +1,11 @@
+package ocudni104.idp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class IdpApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(IdpApplication.class, args);
+    }
+}
