+++ networking-monorepo/services/identity-provider/src/main/java/ocudni104/idp/config/SecurityConfig.java
@@ -1,0 +1,33 @@
+package ocudni104.idp.config;
+
+import ocudni104.idp.federation.FederatedIdentityAuthenticationSuccessHandler;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.annotation.Order;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+public class SecurityConfig {
+
+    @Bean
+    @Order(2)
+    public SecurityFilterChain defaultSecurityFilterChain(HttpSecurity http) throws Exception {
+        http
+            .authorizeHttpRequests(auth -> auth
+                .requestMatchers(
+                    "/.well-known/openid-configuration",
+                    "/oauth2/jwks",
+                    "/actuator/health",
+                    "/error"
+                ).permitAll()
+                .anyRequest().authenticated()
+            )
+            .oauth2Login(oauth2 -> oauth2
+                .successHandler(new FederatedIdentityAuthenticationSuccessHandler())
+            )
+            .formLogin(form -> form.loginPage("/login").permitAll());
+
+        return http.build();
+    }
+}
