+++ networking-monorepo/services/identity-provider/src/main/resources/application.yml
@@ -4,14 +4,6 @@ server:
 spring:
   application:
     name: nm-idp-service
-  cloud:
-    consul:
-      host: ${CONSUL_HOST:localhost}
-      port: ${CONSUL_PORT:8500}
-      discovery:
-        service-name: identity-provider
-        prefer-ip-address: true
-        health-check-path: /actuator/health
   security:
     oauth2:
       client:
