--- networking-monorepo/services/insights-service/src/main/java/com/example/Main.java
@@ -1,26 +1,0 @@
-package com.example;
-
-//TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
-// click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
-public class Main {
-    private final String FEATURE_FLAG ="feature_flag_code";
-    private final String FEATURE_FLAG_2 ="feature_flag_code";
-
-    static void main() {
-        //TIP Press <shortcut actionId="ShowIntentionActions"/> with your caret at the highlighted text
-        // to see how IntelliJ IDEA suggests fixing it.
-        IO.println(String.format("Hello and welcome!"));
-
-        for (int i = 1; i <= 5; i++) {
-            //TIP Press <shortcut actionId="Debug"/> to start debugging your code. We have set one <icon src="AllIcons.Debugger.Db_set_breakpoint"/> breakpoint
-            // for you, but you can always add more by pressing <shortcut actionId="ToggleLineBreakpoint"/>.
-            IO.println("i = " + i);
-        }
-    }
-
-    private void methodThatAlreadyDoesSomething() {
-        // old logic
-        // new fancy logic, securely wrapped
-        // even more old logic
-    }
-}
