--- networking-monorepo/services/insights-service/src/main/java/com/example/insights/InsightsServiceApplication.java
@@ -1,11 +1,0 @@
-package com.example.insights;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class InsightsServiceApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(InsightsServiceApplication.class, args);
-    }
-}
