+++ networking-monorepo/services/insights-service/src/main/java/ocudni104/insights/InsightsServiceApplication.java
@@ -1,0 +1,11 @@
+package ocudni104.insights;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class InsightsServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(InsightsServiceApplication.class, args);
+    }
+}
