+++ networking-monorepo/services/insights-service/src/main/resources/application.yml
@@ -1,0 +1,12 @@
+server:
+  port: 8080
+
+spring:
+  application:
+    name: insights-service
+
+management:
+  endpoints:
+    web:
+      exposure:
+        include: health,info
